Refactor lightning module parameters
